Add challenge client for EdDSA and refactor challenge utils to better support non-native crypto libraries.
